Fetch docs from behaviour callback docs (#84)
* Fetch docs from behaviour callback docs

In some cases it is possible that functions don't have docs, but the callback (or protocol) doc can then be used instead. A common example is Ecto.Repo. When we `use Ecto.Repo`, the functions reside in our module, but the docs should be fetched from the behaviour callback docs.

* Properly handle macro callbacks

* Fix problems with erlang behaviour

* Handle @impl callbacks