Fix autocomplete for many_to_many associations
Using ElixirLS with VS Code, I noticed that schemas with a many_to_many relationship weren't getting autocompletions as I composed ecto queries.

The output was complaining about the `%ManyToMany{}` struct not having a `:related_key`, and I traced it back to here. I've confirmed that this change does fix the issue and allows you to get autocompletions for schemas with `many_to_many` relationships.