Allowed logging of errors to be quieted.
While using elixir-ls, frequently, the code that's sent to the LSP
server (and then to elixir_sense) isn't always syntactically accurate,
and sometimes elixir_sense will fail and spam the logs. After enough
of these spammy logs, emacs throws the error into a mini-buffer to
alert the user to some problem, which is not particularly helpful.

Emitting an error message in these circumstances is dubious, since
there's nothing truly wrong, but I can understand that it might be
helpful for other uses of elixir_ls. To that end, I consolidated the
logging functions into a module and gated them with an application
variable so that other apps can enable or disable logging at their
discretion.