Support finding function definition defined in __using__ macro
This adds heuristic to find the function source location, even if it's
defined insige the `__using__` macro. It supports the following cases:

1. When current module has `use OtherMod` and the function is defined in
`OtherMod`'s `__using__` - it correctly points to a relevant line in
`OtherMod`.

2. When `MyMod` has `use OtherMod` (as above) and `YetAnotherMod` calls
`MyMod.function_defined_in_using`. This basically covers things like
`MyApp.Repo.all`, where `Repo` has `use Ecto.Repo` - now ElixirSense
correctly finds the definition of `all` inside `Ecto.Repo`.

This is not a perfect solution, as it uses regular expressions
and heuristic for find the definition and it might something be wrong.
However, this covers quite important use cases in language servers, so
in my opinion it's worth having, even if it's not 100% precise.